+++ vasttools/source.py
@@ -47,7 +47,6 @@ from typing import List, Tuple, Optional, Union
 
 from radio_beam import Beam
 
-from vasttools import RELEASED_EPOCHS
 from vasttools.survey import Image
 from vasttools.utils import crosshair, filter_selavy_components, read_selavy
 from vasttools.tools import offset_postagestamp_axes
@@ -803,14 +802,13 @@ class Source:
             )
         else:
             selavy_components = read_selavy(row.selavy, cols=[
-                    'island_id',
-                    'ra_deg_cont',
-                    'dec_deg_cont',
-                    'maj_axis',
-                    'min_axis',
-                    'pos_ang'
-                ]
-            )
+                'island_id',
+                'ra_deg_cont',
+                'dec_deg_cont',
+                'maj_axis',
+                'min_axis',
+                'pos_ang'
+            ])
 
         selavy_coords = SkyCoord(
             selavy_components.ra_deg_cont.values,
@@ -841,7 +839,7 @@ class Source:
 
     def show_png_cutout(
         self,
-        epoch: str,
+        index: int,
         selavy: bool = True,
         percentile: float = 99.9,
         zscale: bool = False,
@@ -861,7 +859,7 @@ class Source:
         No access to save.
 
         Args:
-            epoch: The epoch to show.
+            index: Index of the observation to show.
             selavy: If `True` then selavy overlay are shown,
                  defaults to `True`.
             percentile: The value passed to the percentile
@@ -887,7 +885,7 @@ class Source:
         """
 
         fig = self.make_png(
-            epoch,
+            index,
             selavy=selavy,
             percentile=percentile,
             zscale=zscale,
@@ -907,7 +905,7 @@ class Source:
 
     def save_png_cutout(
         self,
-        epoch: str,
+        index: int,
         selavy: bool = True,
         percentile: float = 99.9,
         zscale: bool = False,
@@ -929,7 +927,7 @@ class Source:
         Always save.
 
         Args:
-            epoch: The epoch to show.
+            index: Index of the observation to show.
             selavy: If `True` then selavy overlay are shown,
                  defaults to `True`.
             percentile: The value passed to the percentile
@@ -957,7 +955,7 @@ class Source:
             None
         """
         fig = self.make_png(
-            epoch,
+            index,
             selavy=selavy,
             percentile=percentile,
             zscale=zscale,
@@ -978,49 +976,51 @@ class Source:
 
         return
 
-    def _get_save_name(self, epoch: str, ext: str) -> str:
+    def _get_save_name(self,
+                       index: int,
+                       ext: str
+                       ) -> str:
         """
         Generate name of file to save to.
 
         Args:
-            epoch: Epoch corresponding to requested data
+            index: Index of the requested data
             ext: File extension
 
         Returns:
             Name of file to save.
         """
 
+        row = self.measurements.iloc[index]
+
+        if not ext.startswith("."):
+            ext = f".{ext}"
+
+        source_name = self.name.replace(" ", "_").replace("/", "_")
+        
         if self.pipeline:
-            name_epoch = epoch
+            outfile = f"{source_name}_{index}{ext}"
         else:
-            if "-" in epoch:
-                e_split = epoch.split("-")
-                e = e_split[0]
-                name_epoch = RELEASED_EPOCHS[e] + "-" + e_split[1]
-            else:
-                name_epoch = RELEASED_EPOCHS[epoch]
-        outfile = "{}_EPOCH{}{}".format(
-            self.name.replace(" ", "_").replace(
-                "/", "_"
-            ),
-            name_epoch,
-            ext
-        )
+            field_name = row.field
+            sbid = row.sbid
+
+            outfile = f"{source_name}_{field_name}_SB{sbid}{ext}"
+
         return outfile
 
     def save_fits_cutout(
         self,
-        epoch: str,
+        index: int,
         outfile: Optional[str] = None,
         size: Optional[Angle] = None,
         force: bool = False,
         cutout_data: Optional[pd.DataFrame] = None
     ) -> None:
         """
-        Saves the FITS file cutout of the requested epoch.
+        Saves the FITS file cutout of the requested observation.
 
         Args:
-            epoch: Requested epoch.
+            index: The index of the requested observation.
             outfile: File to save to, defaults to None.
             size: Size of the cutout, defaults to None.
             force: Whether to force the re-fetching
@@ -1032,29 +1032,20 @@ class Source:
             None
 
         Raises:
-            ValueError: If the source does not contain the requested epoch.
+            ValueError: If the source does not contain the requested index.
         """
 
         if (self._cutouts_got is False) or (force):
             if cutout_data is None:
                 self.get_cutout_data(size)
 
-        if epoch not in self.epochs:
-            raise ValueError(
-                "This source does not contain Epoch {}!".format(epoch)
-            )
-
-            return
-
         if outfile is None:
-            outfile = self._get_save_name(epoch, ".fits")
+            outfile = self._get_save_name(index, ".fits")
         if self.outdir != ".":
             outfile = os.path.join(
                 self.outdir,
                 outfile
             )
-
-        index = self.epochs.index(epoch)
 
         if cutout_data is None:
             cutout_row = self.cutout_df.iloc[index]
@@ -1088,7 +1079,8 @@ class Source:
         Returns:
             None
         """
-        self.measurements['epoch'].apply(
+        indices = self.measurements.index.to_series()
+        indices.apply(
             self.write_ann,
             args=(
                 None,
@@ -1117,7 +1109,8 @@ class Source:
             None
         """
 
-        self.measurements['epoch'].apply(
+        indices = self.measurements.index.to_series()
+        indices.apply(
             self.write_reg,
             args=(
                 None,
@@ -1151,8 +1144,8 @@ class Source:
             if cutout_data is None:
                 self.get_cutout_data(size)
 
-        for e in self.measurements['epoch']:
-            self.save_fits_cutout(e, cutout_data=cutout_data)
+        for i in self.measurements.index:
+            self._save_fits_cutout_index(i, cutout_data=cutout_data)
 
     def save_all_png_cutouts(
         self,
@@ -1224,7 +1217,8 @@ class Source:
                 cutout_data=cutout_data
             )
 
-        self.measurements['epoch'].apply(
+        indices = self.measurements.index.to_series()
+        indices.apply(
             self.make_png,
             args=(
                 selavy,
@@ -1511,7 +1505,7 @@ class Source:
 
     def skyview_contour_plot(
         self,
-        epoch: str,
+        index: int,
         survey: str,
         contour_levels: List[float] = [3., 5., 10., 15.],
         percentile: float = 99.9,
@@ -1530,7 +1524,7 @@ class Source:
         the source location and overlays ASKAP contours.
 
         Args:
-            epoch: Epoch requested for the ASKAP contours.
+            index: Index of the requested ASKAP observation.
             survey: Survey requested to be fetched using SkyView.
             contour_levels: Contour levels to plot which are multiples
                  of the local rms, defaults to [3., 5., 10., 15.].
@@ -1555,7 +1549,7 @@ class Source:
             None if save is `True` or the figure object if `False`
 
         Raises:
-            ValueError: If the source does not contain the requested epoch.
+            ValueError: If the index is out of range.
         """
 
         if (self._cutouts_got is False) or (force):
@@ -1563,23 +1557,18 @@ class Source:
 
         size = self._size
 
-        if epoch not in self.epochs:
-            raise ValueError(
-                "This source does not contain Epoch {}!".format(epoch)
-            )
-
+        if index > len(self.measurements):
+            raise ValueError(f"Cannot access {index}th measurement.")
             return
 
         if outfile is None:
-            outfile = self._get_save_name(epoch, ".png")
+            outfile = self._get_save_name(index, ".png")
 
         if self.outdir != ".":
             outfile = os.path.join(
                 self.outdir,
                 outfile
             )
-
-        index = self.epochs.index(epoch)
 
         try:
             paths = SkyView.get_images(
@@ -1631,14 +1620,13 @@ class Source:
         )
 
         if title is None:
-            if self.pipeline:
-                title = "'{}' Epoch {} {}".format(
-                    self.name, epoch, survey
+            obs_time = self.measurements.iloc[index].dateobs
+            title = "{} {}".format(
+                self.name,
+                obs_time.strftime(
+                    "%Y-%m-%d %H:%M:%S"
                 )
-            else:
-                title = "VAST Epoch {} '{}' {}".format(
-                    epoch, self.name, survey
-                )
+            )
 
         ax.set_title(title)
 
@@ -1665,7 +1653,7 @@ class Source:
 
     def make_png(
         self,
-        epoch: str,
+        index: int,
         selavy: bool = True,
         percentile: float = 99.9,
         zscale: bool = False,
@@ -1690,7 +1678,7 @@ class Source:
         Save a PNG of the image postagestamp.
 
         Args:
-            epoch: The requested epoch.
+            index: The index correpsonding to the requested observation.
             selavy: `True` to overlay selavy components, `False` otherwise.
             percentile: The value passed to the percentile
                 normalization function, defaults to 99.9.
@@ -1718,7 +1706,7 @@ class Source:
             force: Whether to force the re-fetching of the cutout data,
                 defaults to `False`.
             disable_autoscaling: Turn off the consistent normalization and
-                calculate the normalizations separately for each epoch,
+                calculate the normalizations separately for each observation,
                 defaults to `False`.
             cutout_data: Pass external cutout_data to be used
                 instead of fetching the data, defaults to None.
@@ -1731,30 +1719,24 @@ class Source:
             None if save is `True` or the figure object if `False`
 
         Raises:
-            ValueError: If the source does not contain the requested epoch.
+            ValueError: If the index is out of range.
         """
 
         if (self._cutouts_got is False) or (force):
             if cutout_data is None:
                 self.get_cutout_data(size)
 
-        if epoch not in self.epochs:
-            raise ValueError(
-                "This source does not contain Epoch {}!".format(epoch)
-            )
-
-            return
+        if index > len(self.measurements):
+            raise ValueError(f"Cannot access {index}th measurement.")
 
         if outfile is None:
-            outfile = self._get_save_name(epoch, ".png")
+            outfile = self._get_save_name(index, ".png")
 
         if self.outdir != ".":
             outfile = os.path.join(
                 self.outdir,
                 outfile
             )
-
-        index = self.epochs.index(epoch)
 
         if cutout_data is None:
             cutout_row = self.cutout_df.iloc[index]
@@ -1927,13 +1909,10 @@ class Source:
             cb.set_label("mJy/beam")
 
         if title is None:
-            epoch_time = self.measurements[
-                self.measurements['epoch'] == epoch
-            ].iloc[0].dateobs
-            title = "{} Epoch {} {}".format(
+            obs_time = self.measurements.iloc[index].dateobs
+            title = "{} {}".format(
                 self.name,
-                epoch,
-                epoch_time.strftime(
+                obs_time.strftime(
                     "%Y-%m-%d %H:%M:%S"
                 )
             )
@@ -2008,7 +1987,7 @@ class Source:
 
     def write_ann(
         self,
-        epoch: str,
+        index: int,
         outfile: str = None,
         crossmatch_overlay: bool = False,
         size: Optional[Angle] = None,
@@ -2020,7 +1999,7 @@ class Source:
         within the image.
 
         Args:
-            epoch: The requested epoch.
+            index: The index correpsonding to the requested observation.
             outfile: Name of the file to write, defaults to None.
             crossmatch_overlay: If True, a circle is added to the
                 annotation file output denoting the crossmatch radius,
@@ -2039,14 +2018,12 @@ class Source:
                 self.get_cutout_data(size)
 
         if outfile is None:
-            outfile = self._get_save_name(epoch, ".ann")
+            outfile = self._get_save_name(index, ".ann")
         if self.outdir != ".":
             outfile = os.path.join(
                 self.outdir,
                 outfile
             )
-
-        index = self.epochs.index(epoch)
 
         neg = False
         with open(outfile, 'w') as f:
@@ -2109,7 +2086,7 @@ class Source:
 
     def write_reg(
         self,
-        epoch: str,
+        index: int,
         outfile: Optional[str] = None,
         crossmatch_overlay: bool = False,
         size: Optional[Angle] = None,
@@ -2120,7 +2097,7 @@ class Source:
         Write a DS9 region file containing all selavy sources within the image
 
         Args:
-            epoch: The requested epoch.
+            index: The index correpsonding to the requested observation.
             outfile: Name of the file to write, defaults to None.
             crossmatch_overlay: If True, a circle is added to the
                 annotation file output denoting the crossmatch radius,
@@ -2139,14 +2116,13 @@ class Source:
                 self.get_cutout_data(size)
 
         if outfile is None:
-            outfile = self._get_save_name(epoch, ".reg")
+            outfile = self._get_save_name(index, ".reg")
         if self.outdir != ".":
             outfile = os.path.join(
                 self.outdir,
                 outfile
             )
 
-        index = self.epochs.index(epoch)
         with open(outfile, 'w') as f:
             f.write("# Region file format: DS9 version 4.0\n")
             f.write("global color=green font=\"helvetica 10 normal\" "